+++ e2e/tests-quill/create_neuron.bash
@@ -11,7 +11,7 @@ teardown() {
 @test "basic create neuron" {
     #account is initialized with 10_000 tokens
     assert_command quill account-balance 345f723e9e619934daac6ae0f4be13a7b0ba57d6a608e511a00fd0ded5866752 --yes --insecure-local-dev-mode
-    assert_string_match 'Balance: 1000000000.00000000 ICP'
+    assert_string_match 'Balance: 1000000000 ICP'
 
     # stake 3 tokens
     assert_command bash -c "quill neuron-stake --amount 3 --name myneur --pem-file \"$PEM_LOCATION/identity.pem\" > stake.call"
@@ -27,7 +27,7 @@ Successfully staked ICP in neuron " #fragment of a correct response
 
     # check that staking worked using get-neuron-info
     assert_command bash -c "quill get-neuron-info \"$NEURON_ID\" --yes --insecure-local-dev-mode"
-    assert_string_match 'Total stake: 3.00000000 ICP'
+    assert_string_match 'Total stake: 3 ICP'
 
     # increase dissolve delay by 6 months
     assert_command bash -c "quill neuron-manage --additional-dissolve-delay-seconds 15778800 --pem-file \"$PEM_LOCATION/identity.pem\" \"$NEURON_ID\" > more-delay.call"