+++ src/commands/neuron_manage.rs
@@ -9,7 +9,7 @@ use candid::{CandidType, Encode, Principal};
 use clap::{Parser, ValueEnum};
 use ic_base_types::PrincipalId;
 use ic_nns_common::pb::v1::{NeuronId, ProposalId};
-use ic_nns_governance::pb::v1::manage_neuron::RefreshVotingPower;
+use ic_nns_governance::pb::v1::manage_neuron::{DisburseMaturity, RefreshVotingPower};
 use ic_nns_governance::pb::v1::{
     manage_neuron::{
         configure::Operation, disburse::Amount, AddHotKey, ChangeAutoStakeMaturity, Command,
@@ -144,6 +144,18 @@ pub struct ManageOpts {
     /// This must be done every so often to avoid neurons diminishing in voting power.
     #[arg(long, alias = "refresh-followers")]
     refresh_following: bool,
+
+    /// Disburse the neuron's maturity to its controller's account.
+    #[arg(long)]
+    disburse_maturity: bool,
+
+    /// Set the percentage of the neuron's maturity to disburse.
+    #[arg(long, value_parser = 1..=100)]
+    disburse_maturity_percentage: Option<i64>,
+
+    /// Disburse the neuron's maturity to the specified NNS account.
+    #[arg(long)]
+    disburse_maturity_to: Option<ParsedNnsAccount>,
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, ValueEnum)]
@@ -155,13 +167,11 @@ enum NativeVisibility {
 pub fn exec(auth: &AuthInfo, opts: ManageOpts) -> AnyhowResult<Vec<IngressWithRequestId>> {
     if opts.ledger {
         ensure!(
-            opts.add_hot_key.is_none() && opts.remove_hot_key.is_none() && !opts.disburse && opts.disburse_amount.is_none() && opts.disburse_to.is_none()
-            && !opts.clear_manage_neuron_followees && !opts.join_community_fund && !opts.leave_community_fund
-            && opts.follow_topic.is_none() && opts.follow_neurons.is_none() && opts.register_vote.is_none() && !opts.reject
-            && opts.set_visibility.is_none() && !opts.refresh_following,
+            !opts.disburse_maturity && opts.disburse_maturity_to.is_none()
+            && opts.disburse_maturity_percentage.is_none(),
             "\
-Cannot use --ledger with these flags. This version of quill only supports the following neuron-manage operations with a Ledger device:
---additional-dissolve-delay-seconds, --start-dissolving, --stop-dissolving, --split, --merge-from-neuron, --spawn, --stake-maturity, --auto-stake-maturity"
+Cannot use --ledger with these flags. This version of quill does not support the --disburse-maturity, --disburse-maturity-to, \
+or --disburse-maturity-percentage flags with a Ledger device"
         );
     }
     let mut msgs = Vec::new();
@@ -427,6 +437,36 @@ Cannot use --ledger with these flags. This version of quill only supports the fo
             command: Some(Command::RefreshVotingPower(RefreshVotingPower {})),
             neuron_id_or_subaccount: id.clone(),
             id: None,
+        })?;
+        msgs.push(args);
+    }
+
+    if opts.disburse_maturity
+        || opts.disburse_maturity_to.is_some()
+        || opts.disburse_maturity_percentage.is_some()
+    {
+        let percentage_to_disburse = opts.disburse_maturity_percentage.unwrap_or(100) as u32;
+        let disburse = match opts.disburse_maturity_to {
+            Some(ParsedNnsAccount::Original(ident)) => DisburseMaturity {
+                percentage_to_disburse,
+                to_account: None,
+                to_account_identifier: Some(ident.into()),
+            },
+            Some(ParsedNnsAccount::Icrc1(account)) => DisburseMaturity {
+                percentage_to_disburse,
+                to_account: Some(account.into()),
+                to_account_identifier: None,
+            },
+            None => DisburseMaturity {
+                percentage_to_disburse,
+                to_account: None,
+                to_account_identifier: None,
+            },
+        };
+        let args = Encode!(&ManageNeuron {
+            id: None,
+            command: Some(Command::DisburseMaturity(disburse)),
+            neuron_id_or_subaccount: id.clone(),
         })?;
         msgs.push(args);
     }